geiq_assessments: migrate AssessmentContractDetailsTab to enums module
So that it can be imported in templates.
